Bill's House color test
Testing of a color improvement method by using a sprite to complement the tileset colors, used in Bill's House where machines overlap the wall.
Sprites needed to be put closer to the entrance and to the center of the room  as to not "pop-in" while scrolling in view. Their visuals have been offset to show in the proper position using my "sprite properties list" system.
The drawing of a sprite is definined by its actual object position and not by its visual location.