Fuchsia City upgrade
Use CopyVideoData to manually write sprites to vram  to bypass Sprite_sets issue, now all Fuchsia pokemons have animated sprites !
One con is a slight lag hiccup when transitioning from Fuchsia to Route 18 or 19.
Removing the gate from Route19 as it lost its purpose and such changes not being in the scope of this mod.