Sprite tearing fix
Fix an issue with sprite tearing.
Code had to by changed in the XY offset portion, instead of doing a second IsInArray check, code was reworked as to be able to continue with the property list from the first check in the Animation table portion.
In this Rework the wPictureID isnt used to store the sprite picture ID as it was originally intended, but to check the status of a sprite being part of the SpecialOAMlist list or not while in the GetSpriteScreenXY function.
wPictureID kept its name because it is still used to store the sprite picture ID in an other part of the code.