Color the Party menu Mons sprites
As described, give color to the sprites of the Mons in the party menu.
The functioninf function load a fixed set of 8 sprites palettes and assign one of them to each Mons.
An alternative non-activated (but functional) function is instead allowing to load 6 palettes based on the Mons in the team, this one allow each mon to have their own custom palette. (As-is it load the ones used for their status/dex image, as a proof of concept).
A con of this one (beyond having to make those 151 palettes)  is that there is a slight but somehow visible delay in the approprioate coloring of the pokemon when moving party member around, how the function work is it load the palettes in the order of the party member, so moving them also mean reloading the palettes.
This second function also recquire a modification to Bankswitch