Implement a system to get color-corrected color palettes at assembly time.
NOT AN INGAME OPTION.
In includes.asm turn "DEF LCD_COLORS EQU 0" to 1 to enable the functionality and play with the options.
All options can be turned OFF independently.
Color matrix : allow to choose 1 of 3 color matrixes to correct the colors
Gamma : apply a gamma value of your choice (the value must be in the 1.0 format, dont omit the decimal even for integers)
White Point : define the max white value, give a hue to the image, can be used to mimic a screen tone, come ;with 3 predefined profiles, more can be added from this file.
For the tinkerers, the rest is in macros\gfx.asm.